requirements.txt
beautifulsoup4==4.9.3
deep-translator==1.7.0
diffbot==2.0.0
emoji-country-flag==1.3.0
Flask==2.0.3
google==3.0.0
huggingface-hub==0.4.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
nltk==3.4.5
numpy==1.22.2
pandas==1.4.1
regex==2022.3.2
requests==2.25.1
scikit-learn==0.24.0
scipy==1.8.0
spacy==3.2.3
tokenizers==0.11.6
torch==1.11.0
tqdm==4.63.0
transformers==4.17.0
urllib3==1.26.8
wasabi==0.9.0
xmltodict==0.12.0
